pyproject.toml
[project]
name = "newsbot"
version = "0.2.0"
description = "News aggregation, LLM briefing, and WhatsApp delivery — powered by Meridiano pipeline"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",

    # Scraping
    "feedparser>=6.0.10",
    "newspaper3k>=0.2.8",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "lxml>=5.1.0",
    "lxml-html-clean>=0.4.0",

    # LLM
    "openai>=1.12.0",

    # Scheduling
    "apscheduler>=3.10.4",

    # Config
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Logging
    "structlog>=24.1.0",

    # Markdown rendering (for web UI)
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "pytest-cov>=4.0.0",
    "httpx",
    "ruff>=0.4.0",
]
prod = [
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "db", "collector", "processor", "delivery", "interactions", "scheduler"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
exclude = ["meridiano-main/*", "*/__pycache__/*", "data/*"]
